refactor: emum moved to mapView folder
