+++ src/app/pages/quickMaps/pages/baselineDetails/mapView/colourGradientType.enum.ts
